Pin all dependencies using pip-tools
This applies the same change we've made in other repos [1]. It's
important to pin the versions of nested dependencies in case new
versions come with unexpected bugs, which may not be caught by CI
if a new version appears at deployment time.

Rather than copy / paste the same old guidance in the README, I've
also created a new Wiki page to centralise it.

[1]: https://github.com/alphagov/notifications-template-preview/commit/dfb6a3f6dec9f7c67501a92c7785dba625374757