Bump NodeJS dependencies to latest versions
Based on running `npm outdated`.

Note: this brings in a version of Sass (aka
dart-ass) past 3, which produces a lot of warnings
about code from GOVUK Frontend. The reasons are
below if you're interested in the detail.

The Sass project moved from using node-sass to
dart-sass as the supported compiler a while ago.
The dart-sass team made a decision to start
deprecating some features supported by node-sass,
one of which is use of the division sign ('/'),
which it is replacing with `math.div()`:

https://sass-lang.com/documentation/breaking-changes/slash-div

GOVUK Frontend uses the division sign in its code.
Moving to `math.div()` would mean it could no
longer support users still using node-sass so it
is still deciding to do this or not. So in the
meantime, running `sass` will emit warnings.

The `sass` command does let you silence these
warnings through the `--quite-deps` flag:

https://sass-lang.com/documentation/cli/dart-sass#quiet-deps

However, we run sass through gulp-sass which, at
the moment, doesn't include that flag in those it
lets you specify (via its options).

So if we did want to silence these warnings, and
maybe detach ourselves from gulp-sass, we'd have
to move to using the `sass` command directly.