Add SkipLink component JS
Version 4 of GOVUK Frontend requires some the JS
for the skip link component to be added to all
pages. See the release notes for details:

https://github.com/alphagov/govuk-frontend/releases/tag/v4.0.0

Section: 'Include JavaScript for skip link to
improve screen reader announcements'.

The approach in these changes includes it as
a separate sub-resource rather than bundling it
with other JS files. It has to load on all pages
so we'd have to bundle it with other JS files on
the pages that have them and also load it on its
own for pages that, before now, didn't have any
JS.

Keeping it separate, and loading it
asynchronously, means it won't block rendering and
will serve from the browser cache on all pages
after the first in a journey.

Also includes changes to the config of the terser
JS compressor to support IE8. It doesn't by
default but this hasn't caught us out until now,
assume because the JS included hasn't used any
language features that needed code that considered
IE8.