Make URLs in alerts clickable
When we send real alerts they might well link to other sources of
information, such as the Flood Warning Service.

In this case it’s clunky to make people have to copy and paste the link
from the content of the URL if they are looking at it on the website.
And if they are coming on the website to check the veracity of the alert
before clicking on the link, then making the link clickable will get
them to the important extra information quicker.